update:: changed the postion of the filters to the hud
